+++ tests/test_quantities.py
@@ -87,12 +87,18 @@ def test_gross_margin(technologies, capacity, market):
     from muse.commodities import is_enduse, is_fuel, is_pollutant
     from muse.quantities import gross_margin
 
+    """
+    Gross margin refers to the calculation
+    .. _here:
+    https://www.investopedia.com/terms/g/grossmargin.asp
+    """
     # we modify the variables to have just the values we want for the testing
-    technologies = technologies.sel(technology=technologies.technology == "soda_shaker")
-    capacity = capacity.sel(asset=capacity.technology == "soda_shaker")
-    capacity[:] = capa = 9
+    selected = capacity.technology.values[0]
 
-    # This will leave 2 environmental outputs and 4 fuel inputs.
+    technologies = technologies.sel(technology=technologies.technology == selected)
+    capa = capacity.where(capacity.technology == selected, drop=True)
+
+    # Filtering commodity outputs
     usage = technologies.comm_usage
 
     technologies.var_par[:] = vp = 2
@@ -104,13 +110,20 @@ def test_gross_margin(technologies, capacity, market):
     market.prices[:] = prices = 3
     market.prices[{"commodity": is_pollutant(usage)}] = env_prices = 6
     # We expect a xr.DataArray with 1 replacement technology
-    actual = gross_margin(technologies, capacity, market.prices)
+    actual = gross_margin(technologies, capa, market.prices)
 
     revenues = prices * prod * sum(is_enduse(usage))
     env_costs = env_prices * envs * sum(is_pollutant(usage))
     cons_costs = prices * fuels * sum(is_fuel(usage))
-    var_costs = vp * (capa**ve) * market.represent_hours / sum(market.represent_hours)
+    var_costs = (
+        vp
+        * ((prod * sum(is_enduse(usage))) ** ve)
+        * market.represent_hours
+        / sum(market.represent_hours)
+    )
+
     expected = revenues - env_costs - cons_costs - var_costs
+    expected *= 100 / revenues
 
     expected, actual = xr.broadcast(expected, actual)
     assert actual.values == approx(expected.values)